Profile: unify /profile and /members/:id behind one component
Both routes now render the same view (icon-headed sections, inline
edit toggles, photo upload, presence). /profile and /members/:id share
the loader, action, and component via members/lib/profile-page.server.ts
and members/components/MemberProfileView.tsx.

Self-view fields (birthday, phone, college ID, dietary restrictions, all
emails, role badges, current-term activity) are now visible on every
member's profile, with inline editing for admins or the member themself.
The activity section shows "My activity" for self and "Activity" for
other viewers; the pending-reviews row only renders when the subject has
unsubmitted reviews. Onboarding gate fires only when the viewer is the
subject.

Drops User.githubUrl (the link is now derived from githubUsername).
Migration: prisma/migrations/20260614120000_drop_user_github_url. Data
loss is limited to user-entered display URLs that the staffing
automation never consumed — the bare username column it relies on is
unchanged.
